backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.8.0
sqlalchemy==2.0.35
alembic==1.12.1
python-dotenv==1.0.0
huggingface-hub==0.23.4
requests==2.31.0
google-generativeai==0.8.3
scikit-learn==1.3.0
easyocr==1.7.1
Pillow==10.0.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
xgboost==1.7.6